### 2018-11-20 #1089 @whedon
[PRE REVIEW]: The Climate Equity Reference Calculator
**Submitting author:** @krueschan (<a href="http://orcid.org/0000-0003-0722-1044">Christian Holz</a>)
**Repository:** <a href="https://github.com/climateequityreferenceproject/cerc-web" target ="_blank">https://github.com/climateequityreferenceproject/cerc-web</a>
**Version:** v3.2.1
