+++ paper/paper.bib
@@ -83,21 +83,21 @@
     place={Manila, London, Cape Town, Washington, et al.},  
     title={Fair Shares: A Civil Society Equity Review of INDCs},  
     url={civilsocietyreview.org/report}, institution={CSO Equity Review Coalition},  
-    author={CSO Equity Review}, year={2015} 
+    author={{CSO Equity Review}}, year={2015} 
 }
 
  @book{CSO_Equity_Review_2016,  
     place={Manila, London, Cape Town, Washington, et al.},  
     title={Setting the Path Towards 1.5°C: A Civil Society Equity Review of Pre-2020 Ambition},  
     url={civilsocietyreview.org/report2016}, institution={CSO Equity Review Coalition},  
-    author={CSO Equity Review}, year={2016} 
+    author={{CSO Equity Review}}, year={2016} 
 }
 
  @book{CSO_Equity_Review_2017,  
     place={Manila, London, Cape Town, Washington, et al.},  
     title={Equity and the Ambition Ratchet: Towards a Meaningful 2018 Facilitative Dialogue},  
     url={civilsocietyreview.org/report2017}, institution={CSO Equity Review Coalition},  
-    author={CSO Equity Review}, year={2017} 
+    author={{CSO Equity Review}}, year={2017} 
 }
 
  @book{Norway_FairShares_2018,  
@@ -132,7 +132,7 @@
     title={Special Report on the Impacts of Global Warming of 1.5°C Above Pre-Industrial Levels and Related Global Greenhouse Gas Emission Pathways, in the Context of Strengthening the Global Response to the Threat of Climate Change, Sustainable Development, and Efforts to Eradicate Poverty},  
     url={http://www.ipcc.ch/report/sr15/},  
     institution={Intergovernmental Panel on Climate Change},  
-    author={IPCC},  
+    author={{IPCC}},  
     year={2018} 
 }
 
@@ -141,7 +141,7 @@
     title={The Climate Equity Reference Calculator Web API},  
     url={https://calculator.climateequityreference.org/api/cerc_web_api.pdf},  
     institution={Stockholm Environmental Institute and EcoEquity},  
-    author={Climate Equity Reference Project},  
+    corpauthor={{Climate Equity Reference Project}},  
     year={2015} 
 }
 
