+++ paper/paper.bib
@@ -141,7 +141,7 @@
     title={The Climate Equity Reference Calculator Web API},  
     url={https://calculator.climateequityreference.org/api/cerc_web_api.pdf},  
     institution={Stockholm Environmental Institute and EcoEquity},  
-    corpauthor={{Climate Equity Reference Project}},  
+    author={{Climate Equity Reference Project}},  
     year={2015} 
 }
 
