### 2018-11-20 #1089 @krueschan
Thanks, whedon, the proofs look good after changes I made.
