### 2018-11-22 #1089 @labarba
@krueschan : I had a quick look at the paper and the repository, and it looks like this submission only covers the web interface of the project. Can you give us your assessment and justification for why the web-interface component alone can qualify as _research software_?
