### 2018-11-22 #1089 @krueschan
@labarba. Thanks for this excellent question. We tried to make this clear in the article, so please let us know if you think this still needs further clarification in the article.

As we write in the article, cerc-web extends the functionality of the "engine" of the calculator quite substantially. In particular, it allows users to compare levels of effort that would be implied by the effort sharing regime based on the settings chosen by the user to reflect their ethical position on one hand, and the level of effort implied by the climate action pledges that countries have actually made on the other. This, in fact, is one of the main ways that the calculator is used in research (about half of the example research pieces that we list at the end have utilized this functionality), and is not available in the "engine" component. (we also mention the API, which also has been used in the past by, albeit now defunct, independent research projects, that arguably would otherwise not have been possible)

Further, users of the calculator typically come from a policy research background and would, in our view, likely not feel comfortable or have the skills to use the calculator without UI. The interactive exploration of implications of different ethical positions and their quantitative representation through the calculator is an important component of research that utilizes the calculator (see Holz, Kartha, Athanasiou 2018), which would not, or only for a very limited user set, be possible without cerc-web.

We also limit the article to the web interface since cerc-web and the calculator "engine" are really quite distinct pieces of software and it would be an odd piece of writing to combine them into one article, besides the "logistical" issues like JOSS, to our knowledge, only supporting one repository link per article etc.

Lastly, we feel that the calculator "engine" is currently sufficiently well-documented (and citeable) through it's technical documentation (Kemp-Benedict at al 2018) and the auto-generated (doxygen) documentation on sourceforge, so we don't think it needs to be covered in a formal publication, while this is not true for the web calculator and its distinct functionality.
