### 2018-11-23 #1089 @jedbrown
Sorry about the mistaken cross-references (too many tabs...). I can edit if needed.

Minor issues I noticed in the paper:
* "authors'" is not used as a possessive, should drop the trailing apostrophe.
* Bib capitalization such as API and CERC.
* It is unusual for authors to "acknowledge" themselves.

The readme has at least one typo ("slite3" -> "sqlite3").

JOSS publications are normally expected to have test suites, but I don't see tests nor documentation for running and extending them.

I also think the paper and/or prominent documentation needs more explanation regarding research use and extensions. Per the [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements):
* The software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)
* The software should be feature complete (no half-baked solutions) and designed for maintainable extension (not one-off modifications). Minor ‘utility’ packages, including ‘thin’ API clients, are not acceptable.
