### 2018-11-23 e4df091
rev 555  (2018-11-23) (krueschan) - expanded research contribution section to paper.md and added relevant paper.bib entries - minor changes to README.md - correcting a few spelling mistakes
