+++ paper/paper.bib
@@ -1,155 +1,195 @@
-@article{Holz_Kartha_Athanasiou_2018, 
-        title={Fairly Sharing 1.5 – National Fair Shares of a 1.5°C-compliant Global Mitigation Effort}, 
-        volume={18}, DOI={10.1007/s10784-017-9371-z}, number={Special Issue: Achieving 1.5°C and Climate Justice}, 
-        journal={International Environmental Agreements: Politics, Law and Economics}, 
-        author={{Holz}, Christian and {Kartha}, Sivan and {Athanasiou}, Tom}, 
-        year={2018}, pages={117–134} 
+@article{Holz_Kartha_Athanasiou_2018,
+    title={Fairly Sharing 1.5 – National Fair Shares of a 1.5°C-compliant Global Mitigation Effort},
+    author={{Holz}, Christian and {Kartha}, Sivan and {Athanasiou}, Tom},
+    volume={18}, DOI={10.1007/s10784-017-9371-z}, number={Special Issue: Achieving 1.5°C and Climate Justice},
+    journal={International Environmental Agreements: Politics, Law and Economics},
+    year={2018}, pages={117–134}
 }
 
 @report{Baer_et_al_2008,
-    location = {Berlin},
     title = {The Greenhouse Development Rights Framework. The Right to Development in a Climate Constrained World, Second Edition},
+    author = {{Baer}, Paul and {Athanasiou}, Tom and {Kartha}, Sivan and {Kemp-Benedict}, Eric},
     url = {http://www.ecoequity.org/docs/TheGDRsFramework.pdf},
     institution = {{Heinrich Böll Foundation}, {Christian Aid}, {EcoEquity} and the {Stockholm Environment Institute}},
-    author = {{Baer}, Paul and {Athanasiou}, Tom and {Kartha}, Sivan and {Kemp-Benedict}, Eric},
+    location = {Berlin},
     date = {2008-11}
 }
 
-@article{Baer_2013, 
-        title={The Greenhouse Development Rights Framework for Global Burden Sharing: Reflection on Principles and Prospects}, 
-        volume={4}, ISSN={17577780}, DOI={10.1002/wcc.201}, number={1}, journal={Wiley Interdisciplinary Reviews: Climate Change}, 
-        author={{Baer}, Paul}, 
-        year={2013}, month={Jan}, pages={61–71} 
+@article{Baer_2013,
+    title={The Greenhouse Development Rights Framework for Global Burden Sharing: Reflection on Principles and Prospects},
+    author={{Baer}, Paul},
+    volume={4}, ISSN={17577780}, DOI={10.1002/wcc.201}, number={1}, journal={Wiley Interdisciplinary Reviews: Climate Change},
+    year={2013}, month={Jan}, pages={61–71}
 }
 
 @software{EKB_Engine,
-    location = {Somerville and Berkeley},
     title = {The Greenhouse Development Rights and Climate Equity Reference Calculator Engine},
+    author = {{Kemp-Benedict}, Eric and {Baer}, Paul and {Kartha}, Sivan and {Athanaisou}, Tom},
     url = {http://gdrs.sourceforge.net},
     publisher = {{Stockholm Environment Institute} and {EcoEquity}},
-    author = {{Kemp-Benedict}, Eric and {Baer}, Paul and {Kartha}, Sivan and {Athanaisou}, Tom},
+    location = {Somerville and Berkeley},
     date = {2014}
 }
 
-@book{EKB_Calculations, 
-    title={Calculations for the Climate Equity Reference Calculator (CERc)}, 
[…]
