+++ paper/paper.md
@@ -24,7 +24,7 @@ affiliations:
    index: 1
  - name: Stockholm Environment Institute US, Somerville
    index: 2
- - name: EcoEquity, Berkeley 
+ - name: EcoEquity, Berkeley
    index: 3
 date: 22 October 2018
 bibliography: paper.bib
@@ -34,7 +34,7 @@ bibliography: paper.bib
 
 In the context of the global response to the threat of anthropogenic climate change, the question arises of how to fairly share the overall global effort required to mitigate climate change. A rich body of literature exists that addresses this question, with several such effort-sharing approaches being proposed, reflecting different ethical positions with regards to fairness and justice. The Climate Equity Reference Project (CERP) framework [@Holz_Kartha_Athanasiou_2018], formerly called Greenhouse Development Rights (GDRs) [@Baer_et_al_2008; @Baer_2013] is such an approach. It takes nations' historical responsibility for greenhouse gas emissions and their financial capacity to reduce emissions and otherwise address climate change as the basis for assigning fair shares of the overall global effort. Importantly, it does so while at the same time protecting the right of the world's poorest to pursue a life free of poverty.
 
-The Climate Equity Reference Calculator and its web interface, ``cerc-web``, allows users to calculate such fair share contributions for each country, based on a user’s selection of equity-related parameters. The ``cerc-web`` interface serves as a user interface to, and enhances the capabilities of the calculator "engine" backend. The "engine," written in C, has been released by the authors' as a separate package [@EKB_Engine]. A technical documentation of the quantitative effort-sharing model that is implemented by the calculator is also available [@EKB_Calculations]. Taken together, the "engine" and ``cerc-web`` represent the only official implementation of the theoretical GDRs and CERP frameworks by their authors, although implementations by other teams exists [@Meinshausen_et_al_2015; @RobiouduPont_et_al_2017; @Pan_etal_2017], often obtaining substantially different results that have not yet been fully explored.
+The Climate Equity Reference Calculator and its web interface, ``cerc-web``, allows users to calculate such fair share contributions for each country, based on a user’s selection of equity-related parameters. The ``cerc-web`` interface serves as a user interface to, and enhances the capabilities of the calculator "engine" backend. The "engine," written in C, has been released by the authors as a separate package [@EKB_Engine]. A technical documentation of the quantitative effort-sharing model that is implemented by the calculator is also available [@EKB_Calculations]. Taken together, the "engine" and ``cerc-web`` represent the only official implementation of the theoretical GDRs and CERP frameworks by their authors, although implementations by other teams exists [@Meinshausen_et_al_2015; @RobiouduPont_et_al_2017; @Pan_etal_2017], often obtaining substantially different results that have not yet been fully explored.
 
 In addition to the calculator "engine," the calculator requires a "core database" as an external component, which contains historical data and future projections for greenhouse gas emissions, GDP, population, Gini coefficients and exchange rates for all countries. The default database, currently in version 7.2, is being frequently updated and is available online, together with its documentation [@Holz_etal_CoreDB_2018]. To our knowledge, at present only a single installation of ``cerc-web`` exists. It is located at https://calculator.climateequityreference.org and is maintained by the authors.
 
@@ -44,13 +44,12 @@ On (i) the Climate Equity Reference framework is an effort sharing _framework_, 
 
[…]

[…]
