+++ paper/paper.md
@@ -46,7 +46,7 @@ On (ii), the calculator "engine" takes as its input an SQLite [@SQLite] database
 
 Finally, on (iii), ``cerc-web`` complements the core calculator "engine" functionality in that it, in its current version, allows end-users to compare countries' actually-articulated emissions reductions pledges with the results of the fair-share calculations based on users' specific ethical position, reflected through their inputs in the user interface. This is achieved through an auxiliary "pledge database" into which maintainers can enter emission reductions pledges as communicated by countries. The authors maintain a published list of their quantifications of these pledges [@Holz_NDCs] and also maintain the pledge database for the ``cerc-web`` instance at https://calculator.climateequityreference.org. ``cerc-web`` also includes a custom user interface for supporting users in managing the pledge database. Finally, for flexibility of use, ``cerc-web`` complements the capabilities of the calculator "engine" by exposing most of the calculator functionality to an API, which can be used by other projects to dynamically incorporate queries to the calculator. The API is comprehensibly documented elsewhere [@CERC_API].
 
-``cerc-web`` and its results have been used extensively in research by the authors and others, most recently in [@Holz_Kartha_Athanasiou_2018; @Norway_FairShares_2018; @ChristianAid_CHOGM_2018; @Richards_Wollenberg_van_Vuuren_2018; @CSO_Equity_Review_2017; @CSO_Equity_Review_2016; @CSO_Equity_Review_2015; @Metcalfe_2015] among other pieces, several of which are cited in the recent Special Report on Global Warming of 1.5°C by the Intergovernmental Panel on Climate Change [@IPCC_SR15], highlighting the research relevance of the software. Especially the feature of ``cerc-web``, described under (iii) above, to allow comparison of countries' climate mitigation pledges with equitable shares of the global effort as calculated by ``cerc-web`` has enabled innovative research. In the context of the mitigation pledges submitted in the context of the Paris Agreement [@PA], enabled the first-ever research report that went beyond a collective assessment of all of these pledges [e.g. @UNFCCC_NDC_Synth; @UNEP_Gap_2015] and assessed the pledges of all countries individually relative to normatively-derived, quantitative benchmarks [@CSO_Equity_Review_2015; @Holz_Kartha_Athanasiou_2018].
+``cerc-web`` and its results have been used extensively in research by the authors and others, most recently in [@Holz_Kartha_Athanasiou_2018; @Norway_FairShares_2018; @ChristianAid_CHOGM_2018; @Richards_Wollenberg_van_Vuuren_2018; @CSO_Equity_Review_2017; @CSO_Equity_Review_2016; @CSO_Equity_Review_2015; @Metcalfe_2015] among other pieces, several of which are cited in the recent Special Report on Global Warming of 1.5°C by the Intergovernmental Panel on Climate Change [@IPCC_SR15], highlighting the research relevance of the software. Especially the feature of ``cerc-web``, described under (iii) above, to allow comparison of countries' climate mitigation pledges with equitable shares of the global effort as calculated by ``cerc-web`` has enabled innovative research. With regards to the mitigation pledges submitted in the context of the Paris Agreement [@PA], this functionality enabled the first-ever research report that went beyond an aggregated, global-level assessment of these pledges [e.g. @UNEP_Gap_2015; @UNFCCC_NDC_Synth] and instead assessed all of these pledges individually relative to normatively-derived, quantitative benchmarks [@CSO_Equity_Review_2015; @Holz_Kartha_Athanasiou_2018].
 
 # Acknowledgements
 
