### 2018-11-23 #1089 @krueschan
@jedbrown, thanks very much for your comments, please find responses below and don't worry about the mistaken cross-references.

>     * "authors'" is not used as a possessive, should drop the trailing apostrophe.
>     * Bib capitalization such as API and CERC.
> The readme has at least one typo ("slite3" -> "sqlite3").

Fixed. Thanks.

>     * It is unusual for authors to "acknowledge" themselves.
We don't name any of the authors in the Acknowledgements section, though one of the individuals in the acknowledgements (Tyler Kemp-Benedict) has the same surname as one of the authors (Eric Kemp-Benedict), so that's perhaps causing confusion here?

> JOSS publications are normally expected to have test suites, but I don't see tests nor documentation for running and extending them.

I have no idea how I would do this for this software. Can you give me a pointer? Or do you think this might not be necessary for this piece of software? In the article, we point to a live installation of the calculator (https://calculator.climateequityreference.org), so perhaps that can serve this purpose?

> I also think the paper and/or prominent documentation needs more explanation regarding research use and extensions. Per the [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements):
>     * The software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)

We believe that the last two paragraphs of the article already covered that point, but have now expanded on the specific research contributions that the software makes and enables, explicitly highlighting where functionality only offered by ``cerc-web`` has been instrumental in research output. paper.md and the pdf article proof have been updated accordingly.

>     * The software should be feature complete (no half-baked solutions) and designed for maintainable extension (not one-off modifications). Minor ‘utility’ packages, including ‘thin’ API clients, are not acceptable.

We strongly believe that this is not a half-baked solution, which is evidenced by the software having been in use for several years and several research projects, both our own as well as others', based on results obtained with it. We are not sure how to further respond to this item, perhaps you can be a bit more specific if there are concerns in this regard?
