### 2018-11-23 #1089 @krueschan
> Testing in headless mode is pretty common practice. https://developer.mozilla.org/en-US/docs/Mozilla/Firefox/Headless_mode or https://developers.google.com/web/updates/2017/06/headless-karma-mocha-chai. If someone is modifying the source code, how do they know it's still producing correct results? (Also, how does a user audit the methods and interpret the results? Are they expected to "trust" the calculator or to reproduce "interesting" findings in their own statistical environment?)

Ok, thanks, I'll look into how to write such tests. I have never done that before - can you perhaps point me to a project that has tests similar to the one that you'd expect in our project so I can see how it's done?

> To help me understand better, of the papers you have cited that use this project, how many involve (a) modifying/extending the code in this repository, (b) running it on different data sets, and/or (c) experimenting with parameters using the calculator? How do those papers explain their methods so as to be reproducible?

Well, all of these papers use our own installation of the project (as we write in the article, to our knowledge ours is the only current installation of the project in existence). (a) However, as these pieces go back to 2015, some of them have been done with older versions of the code, so in that sense they involve different code. (b) Similarly, we update our own "core database" from time to time to reflect the evolving availability of relevant data (e.g. updated GDP projections or greenhouse gas emissions data), so in that sense the older pieces in the list use a different database than the newer ones. Also, as mentioned in the article, I have been using the project with a custom-created database reflecting Canada's provinces, rather than the world's countries, (though that's not cited there because I didn't end up publishing that work, but we could add the conference paper I gave). Further, in one not-yet-published piece, the authors of that piece have created a database reflecting a hypothetical world with three ideal-typical countries to perform sensitivity experiments. (c) Yes, several of the pieces use different parameter sets from each other, as well as several different parameter sets within the same piece. There is another grey lit piece available that could also be added where we perform extensive sensitivity analyses to explorer how sensitive results are to specific parameter choices. Most of the papers cited describe the calculator methodology in general terms; some of the papers include version numbers for the calculator and core database version used.
