### 2018-11-23 #1089 @jedbrown
I'm not immediately familiar with PHP browser testing, but I'd expect it can use similar methods as Python server-side projects like Jupyter notebook.

Okay for now regarding scope; I think we can solicit reviewer feedback. We don't currently have any registered reviewers who have volunteered expertise in PHP. Can you suggest any?
