### 2018-11-24 #1089 @krueschan
@jedbrown. I started to look into writing tests for this, but hoping the review can already proceed before I have finished that task? Your last comment sounds like it can.

There are several people listed for PHP in the [JOSS reviewers list](https://bit.ly/joss-reviewers). Specifically, with PHP as "preferred language" there are @xirdneh, @jaynanavati, @nmullins, and @johnsamuelwrites plus, with PHP as "other language" we've got @linuxscout, @gabrielspadon, @diegoenry, and @MDebasish.
I can also try and find somebody else, but perhaps these folks in the JOSS reviewer pool are an appropriate first step?
