### 2018-12-03 #1089 @danielskatz
@jedbrown , since it seems that you have been doing editor work, I'm going to assign this to you.  If you feel it should be @kthyng instead, please discuss it between you and reassign it if that's what you both decide.
