### 2019-01-09 #1089 @jedbrown
@xirdneh @jaynanavati :wave: Would either of you be available to review this submission from the angle of web development and extensibility for new research purposes? Your review would be in addition to a domain expert who is familiar with web development, but has limited resources to evaluate those aspects of this software. I anticipate some conversation in the review thread to help focus your efforts. Thanks in advance.
