### 2019-01-12 #1089 @jedbrown
@krueschan Would you be willing to add an installation option as a Docker container or similar? This would significantly streamline the review and automate what appears to be a complicated installation process for any users who wish to run their own server (even on localhost to inspect different data, etc.).
