### 2019-01-15 #1089 @krueschan
@jedbrown I have looked into Docker a bit and it seems it could work. alternatively, I'll put together a virtualbox with the software installed for the same purpose.
