### 2019-02-06 #1089 @jedbrown
@krueschan Please let us know if/when a Docker image or VM is available.
