### 2019-02-15 27b8f11
minor changes to paper bibliography
