+++ paper/paper.bib
@@ -82,7 +82,7 @@
  @book{CSO_Equity_Review_2015,
     title={Fair Shares: A Civil Society Equity Review of {INDCs}},
     author={{{CSO} Equity Review}},
-    url={civilsocietyreview.org/report}, institution={{CSO} Equity Review Coalition},
+    url={http://civilsocietyreview.org/report}, institution={{CSO} Equity Review Coalition},
     place={Manila, London, Cape Town, Washington, et al.},
     year={2015}
 }
@@ -90,7 +90,7 @@
  @book{CSO_Equity_Review_2016,
     title={Setting the Path Towards 1.5°C: A Civil Society Equity Review of Pre-2020 Ambition},
     author={{{CSO} Equity Review}},
-    url={civilsocietyreview.org/report2016}, institution={{CSO} Equity Review Coalition},
+    url={http://civilsocietyreview.org/report2016}, institution={{CSO} Equity Review Coalition},
     place={Manila, London, Cape Town, Washington, et al.},
     year={2016}
 }
@@ -98,7 +98,7 @@
  @book{CSO_Equity_Review_2017,
     title={Equity and the Ambition Ratchet: Towards a Meaningful 2018 Facilitative Dialogue},
     author={{{CSO} Equity Review}},
-    url={civilsocietyreview.org/report2017}, institution={{CSO} Equity Review Coalition},
+    url={http://civilsocietyreview.org/report2017}, institution={{CSO} Equity Review Coalition},
     place={Manila, London, Cape Town, Washington, et al.},
     year={2017}
 }
@@ -185,7 +185,7 @@
     year={2015}
 }
 
-@article{Metcalfe_2015, 
+@article{Metcalfe_2015,
     title={Fast, Fair Climate Action Crucial for Health and Equity},
     author={{Metcalfe}, Scott},
     volume={128}, ISSN={1175-8716}, number={1425},
