### 2019-02-20 08cf178
removed old GDRs reference and static css link from calc.js
