### 2019-02-22 f30a52c
minor bug fixes
