### 2019-02-22 ed43f91
added Dockerfile and documentation for dockerized install of cerc-web
