+++ docker/install-devilbox.sh
@@ -1,0 +1,102 @@
+#!/bin/bash
+** This file can be used to install cerc-web into a devilbox docker container
+** from your devilbox directory, run ./shell.sh (or shell.bat) and then
+** execute this file. You might have to create a cerc-web hostname in your
+** hosts file. Refer to devilbox documentation for that. http://devilbox.org
+
+
+mkdir /shared/httpd/cerc-web
+cd /shared/httpd/cerc-web
+git clone https://github.com/climateequityreferenceproject/cerc-web
+ln -s cerc-web/public_html/ htdocs
+
+mysql -u root -h 127.0.0.1 -p  < /shared/httpd/cerc-web/htdocs/databasemanager/pledgedb/db/pledge.sql
+mysql -u root -h 127.0.0.1 -p  < /shared/httpd/cerc-web/htdocs/databasemanager/helpdb/db/help_db.sql
+mysql -u root -h 127.0.0.1 -p -e 'CREATE USER "cerp"@"%" IDENTIFIED BY "cerp";'
+mysql -u root -h 127.0.0.1 -p -e 'GRANT SELECT, INSERT, UPDATE, DELETE, CREATE, DROP, INDEX, ALTER, CREATE TEMPORARY TABLES, CREATE VIEW, EVENT, TRIGGER, SHOW VIEW, CREATE ROUTINE, ALTER ROUTINE, EXECUTE ON `pledges\_cerp`.* TO "cerp"@"%";'
+mysql -u root -h 127.0.0.1 -p -e 'GRANT SELECT, INSERT, UPDATE, DELETE, CREATE, DROP, INDEX, ALTER, CREATE TEMPORARY TABLES, CREATE VIEW, EVENT, TRIGGER, SHOW VIEW, CREATE ROUTINE, ALTER ROUTINE, EXECUTE ON `help\_db`.* TO "cerp"@"%";'
+
+mkdir /shared/httpd/cerc-web/cerc_data
+mkdir /shared/httpd/cerc-web/cerc_data/databases
+mkdir /shared/httpd/cerc-web/cerc_data/sessions
+mkdir /shared/httpd/cerc-web/cerc_data/sessions/user-dbs
+mkdir /shared/httpd/cerc-web/htdocs/tmp/
+
+svn checkout https://svn.code.sf.net/p/gdrs/code/ /shared/httpd/cerc-web/gdrscode
+cd /shared/httpd/cerc-web/gdrscode/gdrsclib
+make CONF=Development clean
+make CONF=Development
+make CONF=Public clean
+make CONF=Public
+
+cd /shared/httpd/cerc-web
+git clone https://github.com/climateequityreferenceproject/cerc-coredb
+
+cd /shared/httpd/cerc-web/
+mkdir PEAR
+mkdir PEAR/HTTP
+mkdir PEAR/temp
+cd PEAR/temp
+git clone https://github.com/pear/pear-core
+git clone https://github.com/pear/HTTP_Request
+git clone https://github.com/pear/Net_Socket
+git clone https://github.com/pear/Net_URL
+mv pear-core/PEAR.php ..
+mv HTTP_Request/Request* ../HTTP
+mv Net_Socket/N* ..
+mv Net_URL/URL.php ../Net
+cd ..
+rm -Rf temp
+pwd
+
+cd /shared/httpd/cerc-web/htdocs/
+rm config.php
+echo '<?php' >> config.php
+echo '$core_db = "/shared/httpd/cerc-web/cerc-coredb/cerc-coredb.sql3";' >> config.php
+echo '$core_db_dev = "/shared/httpd/cerc-web/cerc-coredb/cerc-coredb.sql3";' >> config.php
[…]
