+++ docker/install.sh
@@ -1,0 +1,83 @@
+#!/bin/bash
+** This file can be used to install cerc-web on a LEMP stack
+** it assumes that sqlite3, git, subversion, make, and gcc are installed
+
+
+cd /var/www
+git clone https://github.com/climateequityreferenceproject/cerc-web
+ln -s /var/www/cerc-web/public_html /var/www/html
+
+mysql -u root -h 127.0.0.1 -proot < /var/www/html/databasemanager/pledgedb/db/pledge.sql
+mysql -u root -h 127.0.0.1 -proot < /var/www/html/databasemanager/helpdb/db/help_db.sql
+mysql -u root -h 127.0.0.1 -proot -e 'CREATE USER "cerp"@"%" IDENTIFIED BY "cerp";'
+mysql -u root -h 127.0.0.1 -proot -e 'GRANT SELECT, INSERT, UPDATE, DELETE, CREATE, DROP, INDEX, ALTER, CREATE TEMPORARY TABLES, CREATE VIEW, EVENT, TRIGGER, SHOW VIEW, CREATE ROUTINE, ALTER ROUTINE, EXECUTE ON `pledges\_cerp`.* TO "cerp"@"%";'
+mysql -u root -h 127.0.0.1 -proot -e 'GRANT SELECT, INSERT, UPDATE, DELETE, CREATE, DROP, INDEX, ALTER, CREATE TEMPORARY TABLES, CREATE VIEW, EVENT, TRIGGER, SHOW VIEW, CREATE ROUTINE, ALTER ROUTINE, EXECUTE ON `help\_db`.* TO "cerp"@"%";'
+
+mkdir /var/www/cerc-web/cerc_data
+mkdir /var/www/cerc-web/cerc_data/databases
+mkdir /var/www/cerc-web/cerc_data/sessions
+mkdir /var/www/cerc-web/cerc_data/sessions/user-dbs
+mkdir /var/www/html/tmp/
+
+svn checkout https://svn.code.sf.net/p/gdrs/code/ /var/www/cerc-web/gdrscode
+cd /var/www/cerc-web/gdrscode/gdrsclib
+make CONF=Development clean
+make CONF=Development
+make CONF=Public clean
+make CONF=Public
+
+cd /var/www/cerc-web
+git clone https://github.com/climateequityreferenceproject/cerc-coredb
+
+cd /var/www/html/
+rm config.php
+touch config.php
+chown -R www-data:www-data /var/www
+
+echo '<?php' >> config.php
+echo '$core_db = "/var/www/cerc-web/cerc-coredb/cerc-coredb.sql3";' >> config.php
+echo '$core_db_dev = "/var/www/cerc-web/cerc-coredb/cerc-coredb.sql3";' >> config.php
+echo '$user_db_store = "/var/www/cerc-web/cerc_data/sessions/user-dbs";' >> config.php
+echo '$svg_tmp_dir = "/var/www/html/tmp";' >> config.php
+echo '$param_log_file_name = "/var/www/cerc-web/cerc_data/sessions/param_log.txt";' >> config.php
+echo '$xls_tmp_dir = $user_db_store;' >> config.php
+echo '$xls_file_slug = "cerc_all_output_";' >> config.php
+echo '$xls_copyright_notice = "Climate Equity Reference Project Online Calculator (https://calculator.climateequityreference.org)";' >> config.php
+echo '$calc_engine_path = "/var/www/cerc-web/gdrscode/gdrsclib/dist/Public/GNU-Linux-x86/gdrsclib";' >> config.php
+echo '$calc_engine_path_dev = "/var/www/cerc-web/gdrscode/gdrsclib/dist/Development/GNU-Linux-x86/gdrsclib";' >> config.php
[…]
