+++ docker/README.md
@@ -1,0 +1,12 @@
+# Overview
+This directory contains the Dockerfile used to generate a fully self-contained Docker image of cerc-web and all its dependencies.
+
+# Usage
+1. Obtain Docker for your operating system from https://www.docker.com/
+2. If you use Docker Desktop, you can pull the docker image (climateequityreferenceproject/cerc-web) directly from within Docker Desktop otherwise run on your command line: `docker pull climateequityreferenceproject/cerc-web`.
+3. To start the image from the commandline, type `docker run -i -t -d --name cerc-web -p 80:80  krueschan/cerc-web:latest` (if applicable, replace 'latest' with the specific release you're using, for example "version-3.2.1")
+
+# Finally
+This directory also contains install scripts that can be used to automatically install cerc-web into a devilbox Docker container (http://devilbox.org) or into a standard LEMP stack, such as https://hub.docker.com/r/tonisormisson/dev-lemp
+
+The Docker Hub page of cerc-web is at https://hub.docker.com/r/climateequityreferenceproject/cerc-web.
