### 2019-02-22 9ce88c8
added Dockerfile and documentation for dockerized install of cerc-web
