### 2019-02-22 #1089 @krueschan
@jedbrown I have now created a Docker image. You can fetch it from https://hub.docker.com/r/climateequityreferenceproject/cerc-web but I also have a completely self-contained Dockerfile that's in our project's github repo (sub directory 'docker'), that you could use to `docker build` it yourself.

In the context of creating the Docker image, I have found and fixed a few bugs, so I'd like to change the release version for the article to v3.2.1, how would I do that? Thanks.
