### 2019-02-22 #1089 @jedbrown
Cool! I ran
```
$ docker pull climateequityreferenceproject/cerc-web
$ docker run -it --rm -p 8080:80 climateequityreferenceproject/cerc-web bash
root@c941332d432f:/var/www/cerc-web/public_html# dumb-init /start.sh
```
and went to `localhost:8080`. The first step (starting MySQL) fails if I let it run the default `CMD`; I don't know why. I would suggest using a non-privileged port (such as `-p 8080:80`) in the instructions because that's what a local user would need; usually one would test that way before deploying anyway.
