### 2019-02-23 #1089 @krueschan
Ok, I’ll change the instructions. I can’t reliably reproduce the issue you’re describing (MySQL not starting) except that I that happens to me too sometimes (even with well developed docker images such as devilbox) and restarting the docker service fixes it.

I take it that you managed to run it using the commands you listed in your previous post?
