### 2019-02-23 28d8229
change suggested docker install to port 8080
