+++ public_html/js/calc.js
@@ -3,9 +3,9 @@ var clickedDef = false;
 var popup_size = {width: null, length: null};
 
 $(function() {
-    
+
     init_calc_behavior();
-    
+
     // fieldset show/hide
     $("legend").click(function() {
         if (!clickedDef) {
@@ -19,7 +19,7 @@ $(function() {
             }
         }
     });
-	
+
     $("#country_list_button button").click(function() {
         $('#filterDiv').dialog('open');
     });
@@ -27,16 +27,16 @@ $(function() {
 
     $("legend").siblings().hide();
     $("legend.open").siblings().show();
-	
+
     $('#loading').hide();
-	   
+
     $('#save').append('<button id="spinoff_button" type="button">Copy view to new window</button>');
-    
+
     // If JS is enabled, hide the submit button & show the region-country filter + basic/adv
     $('#submit').hide();
-//    $('#region_country_filter').show(); // country report is now default, 
+//    $('#region_country_filter').show(); // country report is now default,
 //    so by default, region_country_filter should hide
-    
+
     // Region list actions
     $('#regionList').click(changeRegionList);
     $('#btnAdd').click(function () {
@@ -45,12 +45,12 @@ $(function() {
     $('#btnRemove').click(function () {
         moveElement('country_selected','country_available');
     });
-    
+
     // Make sure display options are shown/hidden consistent with the chosen view
     set_display();
-    
+
     $('#loading').html('<img src="img/spinner.gif" alt="loading indicator" />');
-    
+
     //--------------------------------------------------
     // User actions that result in a refresh of the page
     //--------------------------------------------------
@@ -58,7 +58,7 @@ $(function() {
         set_display();
         submit();
     });
-    
+
     $('#table_view').change(function () {
         set_display();
         submit();
@@ -68,45 +68,45 @@ $(function() {
             $('#region_country_filter').show();
         }
     });
-    
+
     // Equity settings panel
     $("a[id|='cbdr']").click(cbdr_grid_select);
-    
+
     // Toggle lux threshold
     $('#do_luxcap, #interp_btwn_thresh').click(lux_thresh_activate);
-    
+
     $('#display_yr').change(submit);
     $('#display_ctry').change(submit);
     $('#display_gases').change(submit);
     $('#decimal_pl').change(submit);
     $('#emergency_path').change(submit);
-    
+
     $('#baseline').change(submit);
     $('#cum_since_yr').change(submit);
     $('#use_lulucf').change(submit);
     $('#use_nonco2').change(submit);
     $('#use_netexports').change(submit);
     $('#r_wt').change(submit);
-    
+
     $('#dev_thresh').change(submit);
     $('#lux_thresh').change(submit);
     $('#do_luxcap').change(submit);
     $('#luxcap_mult').change(submit);
     $('#interp_btwn_thresh').change(submit);
     $('#show_tax_tables').change(submit);
-    
+
     $('#cum_since_yr').change(submit);
     // $('#percent_gwp').change(submit);
     $('#percent_gwp_MITIGATION').change(submit);
     $('#percent_gwp_ADAPTATION').change(submit);
     $('#em_elast').change(submit);
-    
+
     $('#use_kab').change(submit);
     $('#dont_use_kab').change(submit);
     $('#kab_only_ratified').change(submit);
-    
+
     $('#use_mit_lag').change(submit);
-    
+
     $('#use_sequencing').change(submit);
     // TODO: These should only be run if use sequencing is checked
     $('#percent_a1_rdxn').change(submit);
@@ -114,7 +114,7 @@ $(function() {
     $('#end_commitment_period').change(submit);
     $('#a1_smoothing').change(submit);
     $('#mit_gap_borne').change(submit);
-    
+
[…]
