### 2019-02-23 #1089 @jedbrown
Restarting `docker.service` doesn't help for me. I don't understand that problem, but running bash and then issuing the one command isn't an issue for the purposes of reviewing. Yes, it worked and appeared to be fully functional when I ran it like that.
