### 2019-02-23 #1273 @jedbrown
@mtobis :wave: Welcome and thanks for agreeing to review! The comments from @whedon above outline the review process, which takes place in this thread (possibly with issues filed in the project repository). I'll be watching this thread if you have any questions.

I ran the docker image locally via
```
$ docker pull climateequityreferenceproject/cerc-web
$ docker run -it --rm -p 8080:80 climateequityreferenceproject/cerc-web bash
root@c941332d432f:/var/www/cerc-web/public_html# dumb-init /start.sh
```
and pointing my browser to `localhost:8080`, which I think is sufficient to demonstrate that this system (or a derivative thereof) can indeed by hosted by external researchers.
