+++ paper/paper.bib
@@ -36,7 +36,7 @@
     author={{Kemp-Benedict}, Eric and {Athanasiou}, Tom and {Baer}, Paul and {Holz}, Christian and {Kartha}, Sivan},
     abstractNote={The Climate Equity Reference Project ({CERP}) framework allocates responsibility for greenhouse gas mitigation and adaptation based on individual responsibility for historical emissions and capacity to act. The Climate Equity Reference Calculator ({CERc}) is a web-based tool for exploring how specific quantitative expressions of the {CERP} framework affect allocations. This paper documents the calculations that the calculator implements},
     institution={{Stockholm Environment Institute} and {EcoEquity}},
-    url = {http://gdrights.org/wp-content/uploads/2013/02/SEI_TechReport_GDRsCalculations.pdf},
+    DOI={10.5281/zenodo.1748847},
     year={2018}, month={Nov}
 }
 
