### 2019-03-10 #1273 @mtobis
I think the "installation" question is ambiguous in the present case. The end user is not expected to be able to run the code on their own machine, but only to interact with the system through a web interface. Thus I am treating "installation" as meaning "ability to attain functionality", rather than "ability to replicate the service". I think the latter may present some difficulties.

(My attempt to do "docker run" per Jed's message did not work. A server was instantiated but it did not respond.)

I think it would have been better to make clear that the code is in php in the html directory. I spent some time trying to figure out where in my docker image it was and how to get it out!
