### 2019-03-10 #1273 @mtobis
Is this the full paper?

https://github.com/openjournals/joss-papers/blob/joss.01273/joss.01273/10.21105.joss.01273.pdf

I see only a summary and a reference section.
