### 2019-03-10 #1273 @jedbrown
Thanks, @mtobis.

This is a typical length for a JOSS "paper"; it is not meant to duplicate project documentation, which should be sufficiently complete to enable research using the software.

I'm okay with the present state of installation; having reproduced it locally, I think someone who wishes to run/extend the software (e.g., with new data or alternate analysis methods) would be able to get it working. I'm sure the technical issue you experienced with the docker container is resolvable in principle and won't ask you to spend further time on it.
