### 2019-03-10 #1273 @mtobis
This work is essentially a web wrapper around https://sourceforge.net/p/gdrs/code/HEAD/tree/gdrsclib/

Let me state categorically that this is very important work and it should be available and documented.

I am not sure whether that sourceforge tree is considered part of the review. I gather it is not.

I believe my job here is to validate the suitability of this submission for JOSS, not its importance to the world, so unfortunately I hit some snags.

* Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

- I defer to Jed on this, so it's okay.

*  Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

- There is fairly extensive online documentation, but it can't obviously be accessed until some parameters are chosen. It would be better if the help page were the landing page. To my initial exploration no use case examples are provided. This could definitely be improved with modest effort.

*  Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

- Yes, if 1) sourceforge link is considered and 2) the theoretical basis for the calculations is considered sufficient (The actual algorithmic basis is straightforward.)

*  Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

- Assuming this is a review only of the wrapper and not of the underlying calculations, the testing presumably is a rather trivial comparison between a few use cases. This could be automated in principle, but I'd be quite confident in this case with manual tests.

- As a one-off wrapper around a specific code base, scope for contribution is very limited so this is moot. I did not notice a way to report issues or seek support, but again I rather doubt that such issues would arise regarding the wrapper code.

All but one of these issues are not really consequential for the intended uses of the code. I have checked off everything except "example usage".

I strongly recommend altering the code so that a first time user can see the documentation without plunging into parameter choices, and that the online documentation be extended with use cases and user stories.
