### 2019-03-10 #1273 @mtobis
Oh, sorry markdown mangled my bullets. Apologies. The above bulleted items were intended to come in pairs with two bullet styles: alternating between the requirement and my comment.
