### 2019-03-11 4ea0cbc
minor changes to paper (add CSO Review 2018)
