+++ paper/paper.bib
@@ -10,7 +10,7 @@
     title = {The Greenhouse Development Rights Framework. The Right to Development in a Climate Constrained World, Second Edition},
     author = {{Baer}, Paul and {Athanasiou}, Tom and {Kartha}, Sivan and {Kemp-Benedict}, Eric},
     url = {http://www.ecoequity.org/docs/TheGDRsFramework.pdf},
-    institution = {{Heinrich Böll Foundation}, {Christian Aid}, {EcoEquity} and the {Stockholm Environment Institute}},
+    institution = {{Heinrich Böll Foundation}, {Christian Aid}, {EcoEquity} and {Stockholm Environment Institute}},
     location = {Berlin},
     date = {2008-11}
 }
@@ -65,7 +65,7 @@
 }
 
 @book{Holz_etal_CoreDB_2018,
-    title={Climate Equity Reference Calculator Database, version 7.2.0 (Oct 2018), V3},
+    title={Climate Equity Reference Calculator Database, version 7.2.0 ({Oct} 2018), {V3}},
     author={{Holz}, Christian and {Kartha}, Sivan and {Athanasiou}, Tom and {Baer}, Paul and {Kemp-Benedict}, Eric},
     url={http://dx.doi.org/10.7910/DVN/O3H22Z}, DOI={10.7910/DVN/O3H22Z},
     institution={Harvard Dataverse},
@@ -102,6 +102,16 @@
     place={Manila, London, Cape Town, Washington, et al.},
     year={2017}
 }
+
+ @book{CSO_Equity_Review_2018, 
+    place={Manila, London, Cape Town, Washington, et al.},
+    title={After Paris: Inequality, Fair Shares, and the Climate Emergency},
+    url={http://civilsocietyreview.org/report2018},
+    institution={{{CSO} Equity Review Coalition}},
+    author={{{CSO Equity Review}},
+    year={2018}
+}
+
 
  @book{Norway_FairShares_2018,
     title={Norway’s Fair Share of Meeting the Paris Agreement},
