### 2019-03-11 #1273 @krueschan
Thanks very much for the review, @mtobis. Below are responses/further questions:

> I am not sure whether that sourceforge tree is considered part of the review. I gather it is not.

Correct, that is the intention.

> * Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
>   * There is fairly extensive online documentation, but it can't obviously be accessed until some parameters are chosen. It would be better if the help page were the landing page. To my initial exploration no use case examples are provided. This could definitely be improved with modest effort.

The current design, where the "equity splash screen" is shown first that shows the most salient parameters and a sub-set of parameters for them was the result of feedback from end-users who felt overwhelmed if presented with the full complement of settings on first loading the calculator. The settings on the splash screen are documented through extensive glossary entries (I hope the links to glossary entries are identifiable as links, e.g. that clicking "Historical Responsibility" brings up a substantial glossary entry explaining the concept as used here?). So even upon first loading the calculator, the user has access to the documentation of the specific parameter choices presented to them at that stage, which feels perhaps sufficient to me - what else would you expect at this stage?

As far as use cases are concerned, we considered the literature that is references in the manuscript (Adow, Ware, & Viita, 2018; CSO Equity Review, 2015, 2016, 2017, 2018; Holz
et al., 2018; Kartha, Holz, & Athanaisou, 2018; Metcalfe, 2015; Richards, Wollenberg, &
van Vuuren, 2018)

> * Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?
>   * Assuming this is a review only of the wrapper and not of the underlying calculations, the testing presumably is a rather trivial comparison between a few use cases. This could be automated in principle, but I'd be quite confident in this case with manual tests.

Happy to provide manual test cases but I have no idea how I would go about them. Can you point me to an example repository that does that? Or some other relevant documentation?

> * Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>   * As a one-off wrapper around a specific code base, scope for contribution is very limited so this is moot. I did not notice a way to report issues or seek support, but again I rather doubt that such issues would arise regarding the wrapper code.

The repository has a CONTRIBUTING.md file and the calculator itself has a "Please send feedback" link at the bottom of each page.

> I strongly recommend altering the code so that a first time user can see the documentation without plunging into parameter choices, and that the online documentation be extended with use cases and user stories.

Thanks for the suggestion and see my comments above. I am thinking of adding a line of text on top of the "equity splash screen" that points the first time user to a general introduction of our approach, i.e. this page: https://climateequityreference.org/about-the-climate-equity-reference-project-effort-sharing-approach/
This could look like this:
![splash screen with intro](https://climateequityreference.org/files/splash_screen_intro.png)
Do you think that would sufficiently address your concerns?
