### 2019-03-11 #1273 @mtobis
yes that would be an adequate and easy fix
