+++ paper/paper.bib
@@ -103,12 +103,12 @@
     year={2017}
 }
 
- @book{CSO_Equity_Review_2018, 
+ @book{CSO_Equity_Review_2018,
     place={Manila, London, Cape Town, Washington, et al.},
     title={After Paris: Inequality, Fair Shares, and the Climate Emergency},
     url={http://civilsocietyreview.org/report2018},
     institution={{{CSO} Equity Review Coalition}},
-    author={{{CSO Equity Review}},
+    author={{{CSO} Equity Review}},
     year={2018}
 }
 
