### 2019-03-11 d9c92c2
typos and adding DOIs
