+++ paper/paper.bib
@@ -67,7 +67,7 @@
 @book{Holz_etal_CoreDB_2018,
     title={Climate Equity Reference Calculator Database, version 7.2.0 ({Oct} 2018), {V3}},
     author={{Holz}, Christian and {Kartha}, Sivan and {Athanasiou}, Tom and {Baer}, Paul and {Kemp-Benedict}, Eric},
-    url={http://dx.doi.org/10.7910/DVN/O3H22Z}, DOI={10.7910/DVN/O3H22Z},
+    DOI={10.7910/DVN/O3H22Z},
     institution={Harvard Dataverse},
     year={2018}
 }
@@ -75,7 +75,7 @@
  @book{Holz_NDCs,
     title={Estimates of Emissions Levels Associated with the Nationally Determined Contributions ({NDCs})},
     author={{Holz}, Christian and {Athanasiou}, Tom and {Kartha}, Sivan},
-    url={https://doi.org/10.7910/DVN/RIBJXF}, DOI={10.7910/DVN/RIBJXF}, institution={Harvard Dataverse. V3},
+    DOI={10.7910/DVN/RIBJXF}, institution={Harvard Dataverse},
     year={2017}
 }
 
@@ -84,6 +84,7 @@
     author={{{CSO} Equity Review}},
     url={http://civilsocietyreview.org/report}, institution={{CSO} Equity Review Coalition},
     place={Manila, London, Cape Town, Washington, et al.},
+    DOI={10.6084/m9.figshare.5917399},
     year={2015}
 }
 
@@ -92,6 +93,7 @@
     author={{{CSO} Equity Review}},
     url={http://civilsocietyreview.org/report2016}, institution={{CSO} Equity Review Coalition},
     place={Manila, London, Cape Town, Washington, et al.},
+    DOI={10.6084/m9.figshare.5917402},
     year={2016}
 }
 
@@ -100,15 +102,17 @@
     author={{{CSO} Equity Review}},
     url={http://civilsocietyreview.org/report2017}, institution={{CSO} Equity Review Coalition},
     place={Manila, London, Cape Town, Washington, et al.},
+    DOI={10.6084/m9.figshare.5917408},
     year={2017}
 }
 
  @book{CSO_Equity_Review_2018,
     place={Manila, London, Cape Town, Washington, et al.},
-    title={After Paris: Inequality, Fair Shares, and the Climate Emergency},
+    title={After {Paris}: Inequality, Fair Shares, and the Climate Emergency},
     url={http://civilsocietyreview.org/report2018},
     institution={{{CSO} Equity Review Coalition}},
     author={{{CSO} Equity Review}},
+    DOI={10.6084/m9.figshare.7637669},
     year={2018}
 }
 
@@ -150,11 +154,11 @@
 }
 
  @book{CERC_API,
-    title={The Climate Equity Reference Calculator Web {API}},
+    title={The {Climate Equity Reference Calculator} Web {API}},
     author={{Climate Equity Reference Project}},
-    url={https://calculator.climateequityreference.org/api/cerc_web_api.pdf},
     institution={Stockholm Environmental Institute and EcoEquity},
     place={Somerville and Berkeley},
+    DOI={10.5281/zenodo.2590636},
     year={2015}
 }
 
