### 2019-03-11 #1273 @krueschan
> > yes that would be an adequate and easy fix
>
> Ok. @mtobis - I have done this now.

P.S. the textbox is displayed in collapsed state if the user has a 'db' cookie set, which identifies them to not be first time users, so delete this cookie to get the full first time user experience.
