### 2019-03-13 #1273 @krueschan
@mtobis , just checking if, as far as you're concerned, your comments regarding
> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

are now resolved with the recent changes as described above? If so, I think the review process probably expects you to tick the appropriate check box in the checklist on the top of the page, if not, that would be good to know so I can figure out what else to do.
