### 2019-03-13 #1273 @jedbrown
Thanks, @mtobis.

@krueschan It looks like all but the first "missing DOI" is buggy. (This experimental whedon feature needs work.) Can you check that first one?  Can you also check capitalization in references? I think `{Paris Agreement}` needs to be protected in multiple cases.
