### 2019-03-13 4f3e50d
minor changes to paper bibliography
