+++ paper/paper.bib
@@ -7,7 +7,7 @@
 }
 
 @report{Baer_et_al_2008,
-    title = {The Greenhouse Development Rights Framework. The Right to Development in a Climate Constrained World, Second Edition},
+    title = {The {Greenhouse Development Rights Framework}. The Right to Development in a Climate Constrained World, Second Edition},
     author = {{Baer}, Paul and {Athanasiou}, Tom and {Kartha}, Sivan and {Kemp-Benedict}, Eric},
     url = {http://www.ecoequity.org/docs/TheGDRsFramework.pdf},
     institution = {{Heinrich Böll Foundation}, {Christian Aid}, {EcoEquity} and {Stockholm Environment Institute}},
@@ -16,14 +16,14 @@
 }
 
 @article{Baer_2013,
-    title={The Greenhouse Development Rights Framework for Global Burden Sharing: Reflection on Principles and Prospects},
+    title={The {Greenhouse Development Rights Framework} for Global Burden Sharing: Reflection on Principles and Prospects},
     author={{Baer}, Paul},
     volume={4}, ISSN={17577780}, DOI={10.1002/wcc.201}, number={1}, journal={Wiley Interdisciplinary Reviews: Climate Change},
     year={2013}, month={Jan}, pages={61–71}
 }
 
 @software{EKB_Engine,
-    title = {The Greenhouse Development Rights and Climate Equity Reference Calculator Engine},
+    title = {The {Greenhouse Development Rights} and {Climate Equity Reference Calculator} Engine},
     author = {{Kemp-Benedict}, Eric and {Baer}, Paul and {Kartha}, Sivan and {Athanaisou}, Tom},
     url = {http://gdrs.sourceforge.net},
     publisher = {{Stockholm Environment Institute} and {EcoEquity}},
@@ -32,7 +32,7 @@
 }
 
 @book{EKB_Calculations,
-    title={Calculations for the Climate Equity Reference Calculator ({CERc})},
+    title={Calculations for the {Climate Equity Reference Calculator} ({CERc})},
     author={{Kemp-Benedict}, Eric and {Athanasiou}, Tom and {Baer}, Paul and {Holz}, Christian and {Kartha}, Sivan},
     abstractNote={The Climate Equity Reference Project ({CERP}) framework allocates responsibility for greenhouse gas mitigation and adaptation based on individual responsibility for historical emissions and capacity to act. The Climate Equity Reference Calculator ({CERc}) is a web-based tool for exploring how specific quantitative expressions of the {CERP} framework affect allocations. This paper documents the calculations that the calculator implements},
     institution={{Stockholm Environment Institute} and {EcoEquity}},
@@ -49,7 +49,7 @@
 }
 
 @article{RobiouduPont_et_al_2017,
-    title={Equitable Mitigation to Achieve the Paris Agreement Goals},
+    title={Equitable Mitigation to Achieve the {Paris Agreement} Goals},
     author={{Robiou du Pont}, Yann and {Jeffery}, Louise and {Gütschow}, Johannes and {Rogelj}, Joeri and {Christoff}, Peter and {Meinshausen}, Malte},
     volume={7}, DOI={10.1038/nclimate3186},
     journal={Nature Climate Change},
@@ -57,7 +57,7 @@
 }
 
 @article{Pan_etal_2017,
-    title={Exploring Fair and Ambitious Mitigation Contributions Under the Paris Agreement Goals},
+    title={Exploring Fair and Ambitious Mitigation Contributions Under the {Paris Agreement} Goals},
[…]
