### 2019-03-13 #1273 @krueschan
Thanks @jedbrown. That first one is wrong too. Just a book chapter by the same name (and from the same authors) as the book that I am citing, so my citation is fine as is. I have protected the `{Paris Agreement}` as suggested and checked for other examples of capitalization and found a few more. Thanks. Looks like the whedon doi guessing experimental feature is at least as buggy as Mendeley's, though much less intrusive.

Can you also please `@whedon set v3.2.2 as version`, please?
