### 2019-03-13 e41051c
minor changes to paper bibliography
