+++ paper/paper.bib
@@ -65,8 +65,8 @@
 }
 
 @book{Holz_etal_CoreDB_2018,
-    title={{Climate Equity Reference Calculator} Database, version 7.2.0 ({Oct} 2018), {V3}},
-    author={{Holz}, Christian and {Kartha}, Sivan and {Athanasiou}, Tom and {Baer}, Paul and {Kemp-Benedict}, Eric},
+    title={{Climate Equity Reference Calculator} Database, version 7.2.0 ({Sep} 2018), {V4}},
+    author={{Holz}, Christian and {Kartha}, Sivan and {Athanasiou}, Tom},
     DOI={10.7910/DVN/O3H22Z},
     institution={Harvard Dataverse},
     year={2018}
