### 2019-03-13 #1273 @krueschan
found and fixed another minor issue with the bibliography.
