### 2019-03-13 #1273 @jedbrown
@krueschan Can you confirm that Sivan Kartha meets the authorship guidelines with respect to this software? https://joss.readthedocs.io/en/latest/submitting.html#authorship  (I ask because there are no commits authored by him.)

After confirmation, please archive the software at Zenodo or similar and report the DOI here.
