### 2019-03-13 #1273 @krueschan
Yes, I confirm that Sivan absolutely meets the authorship guidelines.
The Zenodo archive DOI is 10.5281/zenodo.2592644
