### 2019-03-13 #1273 @labarba
Small editorial fix in the last paragraph:

“In research by the authors and others, most recently in (Adow … ) among other pieces” >> when you use the citation as part of speech in a sentence, you need to use in-text citation; see:  https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.htm#citation_syntax

Also, comma after “e.g.” in parenthetical a bit later.
