### 2019-03-13 0ccb9ea
minor changes to JOSS paper
