### 2019-03-13 #1273 @krueschan
Ok, good points, thanks.
I fixed it, though using a different approach than you suggested, by joining putting parts of the sentence into the parenthesis with the citations. Also, the comma after e.g.
