### 2019-03-14 #1273 @labarba
Congratulations, @krueschan, your JOSS paper is published!

Big thanks to the editor: @jedbrown, and reviewer: @mtobis 🙏
