### 2023-08-30 #1273 @mtobis
No idea what this is all about. Please leave me off the distribution, at
least until you get it sorted.

mt

On Wed, Aug 30, 2023 at 12:34 PM The Open Journals editorial robot <
***@***.***> wrote:

> ⚠️ Couldn't update published paper. An error happened
> <https://github.com/openjournals/joss-papers/actions/runs/6027584235?check_suite_focus=true>.
> The generated XML metadata file is invalid.
>
> Element issue: [facet 'maxLength'] The value has a length of '50'; this exceeds the allowed maximum length of '32'.
>
