### 2023-08-30 #1273 @jedbrown
@mtobis Sorry about this -- we'll figure out a better workflow for the future. I can't remove you from the issue, but you can click unsubscribe in the right margin.

![image](https://github.com/openjournals/joss-reviews/assets/3303/eda6377e-be39-4239-951c-d4f0b8fe3c68)
